Refactor attendance hours to use BigDecimal and extract DateRange logic
- replaced double fields for totalHours and overtime with BigDecimal in Attendance and AttendanceDto for improved precision.
- extracted date range resolution logic into a new DateRangeResolver component and DateRange record
- adjusted related service logic and tests accordingly.

+++ src/main/java/com/iodsky/motorph/attendance/Attendance.java
@@ -4,6 +4,7 @@ import com.iodsky.motorph.employee.model.Employee;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.UUID;
@@ -34,9 +35,9 @@ public class Attendance {
     private LocalTime timeOut;
 
     @Column(name = "total_hours")
-    private Double totalHours;
+    private BigDecimal totalHours;
 
-    private Double overtime;
+    private BigDecimal overtime;
 
 }
